Added `MarkupStyle` and `StyleCall` classes
`StyleCall` essentially has the same function that `Widget.get_style()`
provided, just as an object instead of an inner function. This is better
because storing the origin method (which has been renamed from `origin`
to `method`) can be done in a linting-accepted manner, and it all looks
nicer.

`MarkupStyle` is basically just `create_markup_style()`, but as an
object. It serves to replace the latter, and build on its foundations.
The biggest advantage of this being a class is that it can store its
original `markup`, and its also just syntactically much neater.